Initial commit:
Began writing classes Board and Knight for Knight's Tour puzzle.
Currenly all points are stored as Numpy arrays, but may have to change
that in future.

Coordinates are in form of (row, col), starting at (0,0) in top left.
Board stores points used in a numpy array. Knight stores moves as an
OrderedDict, with hashed points as keys and a list of possible moves from
each point as values. Points (ndarrays) must be converted to strings
before hashing.