Added animation feature in animate.py
The program wasn't finding a solution in a reasonable amount of time. I
added the animation feature so that I could make sure it wasn't stuck in
an infinite loop, and to see whether I could figure out what else the
problem might be. It turns out the path taken by the knight skips a corner
near the beginning, so it would need to backtrack a considerable number of
steps. Since it searches all paths before backtracking, the running time
will be very long indeed.